+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 ## 3.4.3
 ### Visual changes
-* Added setting for empty label text
+* Added setting to customize legend names for unnamed categories
+* Added setting to toggle today line visibility
+* Added setting to wrap label text
+* Added setting for tooltip data
+* Added random milestone style on initial render
 * Removed the exclusion of empty data from the legend, allowing users to leverage the native "Show items with no data" functionality.
 ### Fixes
 * Fix sorting for the tasks